+++ CeirPannelCode/src/main/java/org/gl/ceir/CeirPannelCode/Model/ConsignmentUpdateRequest.java
@@ -6,7 +6,7 @@ public class ConsignmentUpdateRequest {
 	private String roleType;
 	private int roleTypeUserId;
 	private String txnId;
-	private int userId;
+	private long userId;
 	private String remarks;
 	public int getAction() {
 		return action;
@@ -32,10 +32,11 @@ public class ConsignmentUpdateRequest {
 	public void setTxnId(String txnId) {
 		this.txnId = txnId;
 	}
-	public int getUserId() {
+	
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 	public String getRemarks() {
